# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Finance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Account management
  - name: transactions
    description: Transaction management
  - name: budgets
    description: Budget management
  - name: categories
    description: Category management
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: ["income", "expense", null]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: Month in YYYY-MM format
      responses:
        "200":
          description: List of budgets with spending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetWithSpending"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetBody"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: Month in YYYY-MM format
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending grouped by category
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: Month in YYYY-MM format
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"

  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Get income vs expense trend for last 6 months
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit, investment, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - balance
        - currency
        - color
        - createdAt
        - updatedAt

    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit, investment, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
      required:
        - name
        - type
        - balance

    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit, investment, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        createdAt:
          type: string
      required:
        - id
        - name
        - icon
        - color
        - type
        - createdAt

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
      required:
        - name
        - icon
        - color
        - type

    Transaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - accountId
        - type
        - amount
        - description
        - date
        - createdAt

    CreateTransactionBody:
      type: object
      properties:
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - accountId
        - type
        - amount
        - description
        - date

    UpdateTransactionBody:
      type: object
      properties:
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]

    Budget:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        month:
          type: string
        createdAt:
          type: string
      required:
        - id
        - categoryId
        - amount
        - month
        - createdAt

    BudgetWithSpending:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        month:
          type: string
        spent:
          type: number
        remaining:
          type: number
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        createdAt:
          type: string
      required:
        - id
        - categoryId
        - amount
        - month
        - spent
        - remaining
        - categoryName
        - categoryIcon
        - categoryColor
        - createdAt

    CreateBudgetBody:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        month:
          type: string
      required:
        - categoryId
        - amount
        - month

    UpdateBudgetBody:
      type: object
      properties:
        amount:
          type: number
        month:
          type: string

    DashboardSummary:
      type: object
      properties:
        totalBalance:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        netSavings:
          type: number
        transactionCount:
          type: integer
        budgetUtilization:
          type: number
      required:
        - totalBalance
        - monthlyIncome
        - monthlyExpenses
        - netSavings
        - transactionCount
        - budgetUtilization

    CategorySpending:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        amount:
          type: number
        percentage:
          type: number
      required:
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - amount
        - percentage

    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings